client/requirements.txt
paho-mqtt==1.6.1
pydantic==2.5.2
faker==20.1.0